Utilisation de 'StudentsController' plutôt que 'SearchController'
