+++ app/views/shared/_header.html.haml
@@ -51,7 +51,7 @@
 
             .fr-header__tools
               .fr-header__search
-                = form_with method: :get, url: search_path do |form|
+                = form_with method: :get, url: search_students_path do |form|
                   .fr-search-bar{id: "header-search", role: "search"}
                     = form.label :search, "Recherche", class: "fr-label"
                     = form.text_field :search, placeholder: "Rechercher un élève...", value: @search, class: "fr-input"
