+++ app/views/students/search.html.haml
@@ -3,7 +3,7 @@
 - else
   .fr-table
     %table
-      %caption.fr-h1 #{@students.count} résultats trouvés pour la recherche : #{@search}
+      %caption #{@students.count} résultats trouvés pour la recherche : #{@search}
       %thead
         %th{scope: "col"} Élèves
         %th{scope: "col"} Dernière classe
