+++ features/recherche.feature
@@ -10,7 +10,8 @@ Fonctionnalité: Recherche
   Scénario: Le personnel veut rechercher un élève
     # Le test peut ne pas passer si aucun nom de famille ou prénom contenant un 'A' n'a pas été généré
     Quand je recherche l'élève "a"
-    Alors la page contient "Plusieurs résultats trouvés pour la recherche : A"
+    Alors la page contient "Recherche d'un élève"
+    Et la page contient "résultats trouvés pour la recherche : A"
     Quand je recherche l'élève "XRTEZEDE"
     Alors la page contient "Aucun résultat trouvé pour la recherche : XRTEZEDE"
     Quand je recherche l'élève "juan francois mich"
