+++ spec/requests/students_controller_spec.rb
@@ -36,4 +36,12 @@ RSpec.describe StudentsController do
       it { is_expected.to render_template(:show) }
     end
   end
+
+  describe "#unify" do
+    let(:controller) { described_class.new }
+
+    it "returns transformed string" do
+      expect(controller.send(:unify, "Jüãn-Frânçois Mîchäèl")).to eq "JUAN FRANCOIS MICHAEL"
+    end
+  end
 end
